requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
smmap==5.0.2
Werkzeug==3.1.3
